requirements.txt
streamlit>=1.38.0
google-generativeai>=0.8.3
pandas>=2.2.2
numpy>=1.26.4
plotly>=5.22.0
python-dotenv>=1.0.1
matplotlib>=3.9.2
seaborn>=0.13.2


